+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 
 ## 2.1.6-dev
 
-- [ ] 新功能：新增意向 `ChangeDefinitionInjectionModeIntention` / NEW: New intention `ChangeDefinitionInjectionModeIntention`
+- [x] 新功能：新增意向 `ChangeDefinitionInjectionModeIntention` / NEW: New intention `ChangeDefinitionInjectionModeIntention`
 - [ ] 性能分析与优化 / Performance analysis and optimization
 - [ ] 完善参考文档 / Improve reference documentation
 - [x] 其他优化与BUG修复 / Other optimizations and bug fixes